feat(hermes): scaffold new architecture
